# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Integración de Empleados",
    "description": "API para gestión de empleados, solicitudes y jerarquía",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/onboarding/estadisticas/resumen-general": {
      "get": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Get Resumen General Rendimiento",
        "operationId": "get_resumen_general_rendimiento_v1_onboarding_estadisticas_resumen_general_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fecha_desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Desde"
            }
          },
          {
            "name": "fecha_hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Hasta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumenRendimientoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/estadisticas/desempenio-por-destinatario": {
      "get": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Get Desempenio Por Destinatario",
        "operationId": "get_desempenio_por_destinatario_v1_onboarding_estadisticas_desempenio_por_destinatario_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fecha_desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Desde"
            }
          },
          {
            "name": "fecha_hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Hasta"
            }
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Top N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DesempenioDestinatarioResponse"
                  },
                  "title": "Response Get Desempenio Por Destinatario V1 Onboarding Estadisticas Desempenio Por Destinatario Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/estadisticas/desempenio-por-solicitud": {
      "get": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Get Desempenio Por Tipo Solicitud",
        "operationId": "get_desempenio_por_tipo_solicitud_v1_onboarding_estadisticas_desempenio_por_solicitud_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fecha_desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Desde"
            }
          },
          {
            "name": "fecha_hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Hasta"
            }
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Top N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DesempenioTipoSolicitudResponse"
                  },
                  "title": "Response Get Desempenio Por Tipo Solicitud V1 Onboarding Estadisticas Desempenio Por Solicitud Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/onboarding/estadisticas/tiempo-promedio-por-estado": {
      "get": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Get Tiempo Promedio Por Estado",
        "operationId": "get_tiempo_promedio_por_estado_v1_onboarding_estadisticas_tiempo_promedio_por_estado_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fecha_desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Desde"
            }
          },
          {
            "name": "fecha_hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Hasta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EstadoRendimientoResponse"
                  },
                  "title": "Response Get Tiempo Promedio Por Estado V1 Onboarding Estadisticas Tiempo Promedio Por Estado Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumenRendimientoResponse": {
        "properties": {
          "total_solicitudes": {
            "type": "integer",
            "title": "Total Solicitudes"
          },
          "solicitudes_finalizadas": {
            "type": "integer",
            "title": "Solicitudes Finalizadas"
          },
          "solicitudes_rechazadas": {
            "type": "integer",
            "title": "Solicitudes Rechazadas"
          },
          "solicitudes_activas": {
            "type": "integer",
            "title": "Solicitudes Activas"
          },
          "tasa_cierre": {
            "type": "number",
            "title": "Tasa Cierre"
          },
          "tiempo_promedio_total_minutos": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tiempo Promedio Total Minutos"
          },
          "percentil_50_total_minutos": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentil 50 Total Minutos"
          },
          "percentil_90_total_minutos": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentil 90 Total Minutos"
          },
          "estado_mas_lento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado Mas Lento"
          },
          "por_estado": {
            "items": {
              "$ref": "#/components/schemas/EstadoRendimientoResponse"
            },
            "type": "array",
            "title": "Por Estado"
          },
          "top_destinatarios_lentos": {
            "items": {
              "$ref": "#/components/schemas/TopRendimientoItemResponse"
            },
            "type": "array",
            "title": "Top Destinatarios Lentos"
          },
          "top_tipos_solicitud_lentos": {
            "items": {
              "$ref": "#/components/schemas/TopRendimientoItemResponse"
            },
            "type": "array",
            "title": "Top Tipos Solicitud Lentos"
          }
        },
        "type": "object",
        "required": [
          "total_solicitudes",
          "solicitudes_finalizadas",
          "solicitudes_rechazadas",
          "solicitudes_activas",
          "tasa_cierre",
          "por_estado",
          "top_destinatarios_lentos",
          "top_tipos_solicitud_lentos"
        ],
        "title": "ResumenRendimientoResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}